refactor pike.model into pike.smb2 package
extensibility is the goal. with this refactoring, there
are no functional changes. the only code changes have been
to bring in pep8 conventions and fix minor bugs on sight

from here, it will be easier to separate the connection/transport
from the protocol handling code, which should give better
control to subclasses to decide how data gets serialized and sent
on the socket