pike.path: use Tree as _drv
instead of overriding __new__ and causing problems for myself, use the
existing _drv slot of the PureWindowsPath to store the Tree object.

Inherited operations already consider and pass around the _drv, so there
is no need for custom carrying logic. The custom logic is only for
decoding the channel and tree objects from the _drv.

Additionally implement `join_from_root` to "rehome" an existing PikePath
onto a new Tree.